Volume distribution and Momentum compensation
The particles now distrubute throughout the volume of the radius, not the surface area. As each particle is supposed to be able to travel the maximum distance to the center to be the radius, momentum compensation is being implemented, to give particles less than the radius distance away the corresponding momentums they'd have if released from the radius and allowed to reach that position.
Momentum Compensation is being worked on, all other math functions as intented.