Seperated Calculation and Visualization Components
MultObjects.py has begun its transformation into two seperate files for speed optimization. One file for just pure caluclation and position caching, and the other for visualization. (To Be Constructed as of this message).
ISSUE- Still relying on the vpython vector implementation due to the complexities of emulating it elsewhere..